capstone/backend/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
djangorestframework==3.12.4
idna==3.2
mysql-connector==2.2.9
mysql-connector-python==8.0.26
mysqlclient==2.0.3
numpy==1.21.2
oauthlib==3.1.1
pandas==1.3.3
protobuf==3.17.3
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2s
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
tablib==0.12.1
